refactor(dev): unify frontend-bundle-visualizer-report.mjs and frontend-js-size.mjs
